Refactor to make code less confusing.
